refactor: test predictor inside model workflow test
